% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-07-calculate_coalition_probs.R
\name{calculate_coalition_probs}
\alias{calculate_coalition_probs}
\title{Based on the modeling results per party, compute probabilities of coalitions}
\usage{
calculate_coalition_probs(modelResults, dataPrep, predDate)
}
\arguments{
\item{modelResults}{Results of the stan model}

\item{dataPrep}{prepared poll data as returned by preparePollData}

\item{predDate}{when the predictions are created}
}
\description{
Based on the modeling results per party, compute probabilities of coalitions
}
