% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01b-checkPollData.R
\name{checkForNewPolls}
\alias{checkForNewPolls}
\title{Check if there are new polls}
\usage{
checkForNewPolls(latestPolls, latestPollsDb)
}
\arguments{
\item{latestPolls}{data.frame with columns "Institut" and "latestPoll" as
returned from latestPollPerPollster()}

\item{latestPollsDb}{data.frame with columns "Institut" and "latestPollDb" as
returned from getLatestPollsDb()}
}
\value{
logical
}
\description{
Check if there is a newer poll available on the web than in the database
for any pollster.
}
