% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01b-checkPollData.R
\name{checkIfPollDataChanged}
\alias{checkIfPollDataChanged}
\title{Check if data was changed}
\usage{
checkIfPollDataChanged(polls, cred)
}
\arguments{
\item{polls}{data.frame scraped from wahlrecht.de. Reaches back a whole year.}

\item{cred}{credentials specified in scriptMulti.R}
}
\value{
logical
}
\description{
Check if old poll data was changed after a pollster published new data.
}
