% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-04-modelling.R
\name{compileRunModel}
\alias{compileRunModel}
\title{Run stan model}
\usage{
compileRunModel(
  data,
  dev = FALSE,
  iter = ifelse(dev, 20, 125),
  warmup = ifelse(dev, 10, 275),
  control = list(max_treedepth = ifelse(dev, 3, 13), adapt_delta = 0.8),
  chains = 8,
  parallel_chains = 8,
  seed = 12345
)
}
\arguments{
\item{data}{data.frame}

\item{dev}{logical Estimate a very bad, but fast model during code development
or in tests}

\item{iter}{iterations}

\item{warmup}{warmup iterations}

\item{control}{list with additional control arguments}

\item{chains}{chains}

\item{parallel_chains}{number of cores on which chains are calculated}

\item{seed}{seed}
}
\description{
Run stan model
}
