% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-05-plotElectionData.R
\name{dropPredsBeforeFirstPoll}
\alias{dropPredsBeforeFirstPoll}
\title{Drop predictions before the first poll of a party}
\usage{
dropPredsBeforeFirstPoll(dat, pollData)
}
\arguments{
\item{dat}{data.frame Data to be filtered, must contain columns "party" and
"time"}

\item{pollData}{data.frame Poll data, must contain columns for all parties
and "Datum"}
}
\description{
For new parties, we don't want backwards predictions for the time before the
party was founded. Therefore we keep only predictions for the date of the first
poll with this party or later.
}
