% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-06-forecastTable.R
\name{eventsDE}
\alias{eventsDE}
\title{Calculate probs for events}
\usage{
eventsDE(modelResults, data, predDate)
}
\arguments{
\item{modelResults}{model results}

\item{data}{data}

\item{predDate}{prediction date}
}
\description{
Calculate probs for events
}
