% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtils.R
\name{fillDateGapsPerEvent}
\alias{fillDateGapsPerEvent}
\title{Fill up gaps between dates with data points from the most recent event}
\usage{
fillDateGapsPerEvent(df)
}
\arguments{
\item{df}{Incomplete dataframe with percent estimates for events}
}
\value{
Time-consistent dataframe
}
\description{
All date gaps are filled up. This translates into the creation
of new rows for each coalition event and a percent estimation value equal to
the most recent
}
