% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal-01-get_election_results_from_url.R
\name{get_election_results_from_url}
\alias{get_election_results_from_url}
\title{Queries election results from url}
\usage{
get_election_results_from_url(urls)
}
\arguments{
\item{urls}{url where the election results can be found}
}
\value{
dataframe with election results
}
\description{
Queries election results from url
}
