% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal-03-get_latest_governing_coalitions.R
\name{get_latest_gov_coalitions}
\alias{get_latest_gov_coalitions}
\title{Load new coalition data from wikipedia}
\usage{
get_latest_gov_coalitions()
}
\value{
dataframe with variables land, governing_coalition, year
}
\description{
Loads the parties of the latest governing coalition for the federal states of
Germany from wikipedia.
}
