% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtils.R
\name{getDateIdInTimeseq}
\alias{getDateIdInTimeseq}
\title{Get index of a date}
\usage{
getDateIdInTimeseq(timeSeq, date)
}
\arguments{
\item{timeSeq}{Vector of numbers}

\item{date}{a date in the format "YYYY-mm-dd"}
}
\value{
numeric, an index
}
\description{
Get the index of the entry of a sequence that is equal to the number of weeks
between 1970-01-04 and the given date
}
