% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-06-forecastTable.R
\name{getFactSurvey}
\alias{getFactSurvey}
\title{Prepare survey data}
\usage{
getFactSurvey(pollData)
}
\arguments{
\item{pollData}{data frame with survey data}
}
\description{
Prepare survey data
}
