% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-06-forecastTable.R
\name{getForecastAllTable}
\alias{getForecastAllTable}
\title{Prepare table with election forecast}
\usage{
getForecastAllTable(forecastData)
}
\arguments{
\item{forecastData}{data with forecast}
}
\description{
Prepare table with election forecast
}
