% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01b-checkPollData.R
\name{getLatestPollsDb}
\alias{getLatestPollsDb}
\title{Get date of latest poll per pollster in the database}
\usage{
getLatestPollsDb(cred)
}
\arguments{
\item{cred}{DB credentials}
}
\value{
data.frame with columns "Institut" and "latestPollDb"
}
\description{
Get date of latest poll per pollster in the database
}
