% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01b-checkPollData.R
\name{latestPollPerPollster}
\alias{latestPollPerPollster}
\title{Date of latest poll per pollster}
\usage{
latestPollPerPollster(fact_survey)
}
\arguments{
\item{fact_survey}{data.frame with columns "Institut" and "Datum"}
}
\description{
Date of latest poll per pollster
}
