% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partyUtils.R
\name{mapPartyNamesToNumbers}
\alias{mapPartyNamesToNumbers}
\title{Map party names to numeric IDs
Will throw an error when facing a given unknown party}
\usage{
mapPartyNamesToNumbers(parties)
}
\arguments{
\item{parties}{actual data of parties}
}
\description{
Map party names to numeric IDs
Will throw an error when facing a given unknown party
}
