% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-06-forecastTable.R
\name{partOfGovernmentDE}
\alias{partOfGovernmentDE}
\title{Calculate part of government probabilities}
\usage{
partOfGovernmentDE(fact_coalition_prob, predDate)
}
\arguments{
\item{fact_coalition_prob}{coalition probs}

\item{predDate}{prediction date}
}
\description{
Calculate part of government probabilities
}
