% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-05-plotElectionData.R
\name{plotElectionData}
\alias{plotElectionData}
\title{Plot election forecast}
\usage{
plotElectionData(
  modelResults,
  data,
  predDate,
  pollData,
  start = "2016-01-01",
  end = data$nextElectionDate
)
}
\arguments{
\item{modelResults}{model results}

\item{data}{data}

\item{predDate}{prediction date}

\item{pollData}{poll data}

\item{start}{start date}

\item{end}{end date}
}
\description{
Plot election forecast
}
