% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-06-forecastTable.R
\name{prepareForecastTable}
\alias{prepareForecastTable}
\title{Prepare table with election forecast}
\usage{
prepareForecastTable(modelResults, data, predDate)
}
\arguments{
\item{modelResults}{model results}

\item{data}{data}

\item{predDate}{prediction date}
}
\description{
Prepare table with election forecast
}
