% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-02-preparePollData.R
\name{preparePollData}
\alias{preparePollData}
\title{Prepare poll data}
\usage{
preparePollData(pollData, elections, predDate, minDate = predDate - 365 * 16)
}
\arguments{
\item{pollData}{poll data}

\item{elections}{election data}

\item{predDate}{prediction date}

\item{minDate}{minimum date}
}
\description{
Prepare poll data
}
