% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01a-getPollData.R
\name{printLatestPollPerPollster}
\alias{printLatestPollPerPollster}
\title{Print date of latest poll per pollster}
\usage{
printLatestPollPerPollster(fact_survey)
}
\arguments{
\item{fact_survey}{data.frame with columns "Institut" and "Datum"}
}
\description{
Print date of latest poll per pollster
}
