% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_cmdstan_samples_coalition_model.R
\name{rearrange_cmdstan_samples_coalition_model}
\alias{rearrange_cmdstan_samples_coalition_model}
\title{rearrange_cmdstan_samples_coalition_model}
\usage{
rearrange_cmdstan_samples_coalition_model(samples_cmdstan, model_data_list)
}
\arguments{
\item{samples_cmdstan}{samples generated by cmdstan}

\item{model_data_list}{list of model data containing entries for K1, K2, K3, N_new}
}
\value{
list of rearranged samples
}
\description{
Rearrange the samples from the coalition model so they can be used in the downstream code
}
