% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal-01-get_election_results_from_url.R
\name{split_election_results_by_year}
\alias{split_election_results_by_year}
\title{Splits election results by year}
\usage{
split_election_results_by_year(election_results)
}
\arguments{
\item{election_results}{list containing election results, each element is a
dataframe containing a column 'year'}
}
\value{
list with dataframes
}
\description{
Splits election results by year
}
