% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes-01b-checkPollData.R
\name{stopIfNewPollster}
\alias{stopIfNewPollster}
\title{Throw error if there are new pollsters}
\usage{
stopIfNewPollster(pollstersWeb, pollstersDb)
}
\arguments{
\item{pollstersWeb}{character vector with pollsters from wahlrecht.de}

\item{pollstersDb}{character vector with pollsters form db}
}
\description{
If www.wahlrecht.de delivers a pollster (institute) that is not known in the
db, this leads to empty institute values in the database and some data will
be ignored. This function throws an error if there is a new institute.
}
