% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtils.R
\name{weeksBetweenOriginAndDate}
\alias{weeksBetweenOriginAndDate}
\title{Weeks between 1970-01-04 and a given date}
\usage{
weeksBetweenOriginAndDate(date)
}
\arguments{
\item{date}{a date in the format "YYYY-mm-dd"}
}
\value{
numeric
}
\description{
Compute the number of weeks between 1970-01-04 and the given date
Only completed weeks count; e.g., if it's 1 week and 6 days, the result is
1 (week)
}
